refactor: domain 하위 패키지 exception 추가
