+++ src/main/java/subway/application/service/StationService.java
@@ -2,7 +2,7 @@ package subway.application.service;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import subway.domain.NotFoundStationException;
+import subway.domain.exception.NotFoundStationException;
 import subway.domain.Station;
 import subway.infrastructor.repository.StationJpaEntity;
 import subway.infrastructor.repository.StationRepository;
