+++ src/main/java/subway/domain/exception/NotDownStationWhenDeleteSectionException.java
@@ -1,4 +1,4 @@
-package subway.domain;
+package subway.domain.exception;
 
 public class NotDownStationWhenDeleteSectionException extends RuntimeException {
