+++ src/main/java/subway/domain/exception/NotEqualUpStationAndDownStationException.java
@@ -1,4 +1,4 @@
-package subway.domain;
+package subway.domain.exception;
 
 public class NotEqualUpStationAndDownStationException extends RuntimeException {
