+++ src/main/java/subway/domain/exception/NotFoundLineException.java
@@ -1,4 +1,4 @@
-package subway.domain;
+package subway.domain.exception;
 
 public class NotFoundLineException extends RuntimeException {
