+++ src/main/java/subway/domain/exception/NotFoundSectionException.java
@@ -1,4 +1,4 @@
-package subway.domain;
+package subway.domain.exception;
 
 public class NotFoundSectionException extends RuntimeException {
