+++ src/main/java/subway/domain/exception/NotFoundStationException.java
@@ -1,4 +1,4 @@
-package subway.domain;
+package subway.domain.exception;
 
 public class NotFoundStationException extends RuntimeException {
