+++ src/main/java/subway/infrastructor/repository/SectionPersistenceRepository.java
@@ -5,8 +5,8 @@ import subway.application.service.output.LineLoadRepository;
 import subway.application.service.output.SectionCommandRepository;
 import subway.application.service.output.SectionLoadRepository;
 import subway.domain.Line;
-import subway.domain.NotFoundSectionException;
-import subway.domain.NotFoundStationException;
+import subway.domain.exception.NotFoundSectionException;
+import subway.domain.exception.NotFoundStationException;
 import subway.domain.Section;
 
 import java.util.ArrayList;
